% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybootpval.R
\name{bootpval}
\alias{bootpval}
\title{Bootstrap p-value Calculation}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{A numeric vector containing the sample data.}

\item{conf.level}{The confidence level for the confidence interval (default is 0.95).}

\item{iter}{The number of bootstrap iterations (default is 3000).}

\item{mu0}{The null hypothesis value for the population mean (default is 0).}

\item{test}{A character string indicating the type of test. Options are:
- "two" for a two-sided test,
- "upper" for an upper one-sided test,
- "lower" for a lower one-sided test (default is "two").}
}
\value{
A list containing:
        \item{pvalue}{The calculated p-value based on the bootstrap resampling.}
        \item{tcalc}{The calculated t-statistic for the observed data.}
        \item{n}{The sample size of the data.}
        \item{x}{The input sample data.}
        \item{test}{The test type used ("two", "upper", or "lower").}
        \item{ci}{The confidence interval for the resampled means.}
}
\description{
This function calculates the p-value based on bootstrapping for a given sample.
It allows for one-sided or two-sided hypothesis testing and provides a confidence interval
for the mean difference based on bootstrap resampling.
}
\examples{
# Example usage for two-sided test
bootpval(x = rnorm(30), conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")

# Example usage for upper one-sided test
bootpval(x = rnorm(30), conf.level = 0.95, iter = 3000, mu0 = 0, test = "upper")

# Example usage for lower one-sided test
bootpval(x = rnorm(30), conf.level = 0.95, iter = 3000, mu0 = 0, test = "lower")
}
