% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab7Task7.R
\docType{data}
\name{fire}
\alias{fire}
\title{DDT data set.}
\format{
## `who`
A data frame with 144 rows and 5 variables:
\describe{
  \item{distance}{distance of fire damage, unknown units}
  \item{damage}{damage from the fire, unknown units}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
fire
}
\description{
A data set collected by the US forces from the tennessee river and its tributaries
}
\keyword{datasets}
