% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialex.R
\name{mybin}
\alias{mybin}
\title{Binomial Experiment}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{Number of iterations (the number of binomial experiments to run)}

\item{n}{Number of trials per binomial experiment}

\item{p}{Probability of success for each trial in the binomial experiment}
}
\value{
A vector containing the number of successes in each binomial experiment.
}
\description{
Binomial Experiment
}
\examples{
mybin(iter = 1000, n = 20, p = 0.5)
}
