% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltpd.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem for Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Integer. The sample size for each iteration.}

\item{iter}{Integer. The number of iterations to simulate.}

\item{lambda}{Numeric. The rate (mean) parameter for the Poisson distribution. Default is 10.}

\item{...}{Additional arguments to pass to the histogram function.}
}
\value{
A layout with three plots:
}
\description{
This function simulates the Central Limit Theorem by drawing random samples from a Poisson distribution, calculating the sample means, and visualizing their distribution using histograms. It also generates a barplot for the sample values and compares the histogram of sample means with the theoretical normal distribution.
}
\examples{
# Simulate with sample size 10, 1000 iterations, and lambda = 10
mycltp(n = 10, iter = 1000, lambda = 10)

}
