% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeo2.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric Experiment Simulation}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{Integer, how many samplings are repeated, default 100}

\item{N}{Integer, total size of population or total number of items in a population, default 20}

\item{r}{Integer, number of considered successes in the population, default 12}

\item{n}{Integer, sample size, how many are drawn in each iteration, default 5}
}
\value{
A table of proportions of successes from the hypergeometric distribution simulation.
}
\description{
Hypergeometric Experiment Simulation
}
\examples{
myhyper(iter = 1000, N = 20, r = 12, n = 5)
}
