% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A function that calculates the log-likelihood for given data and parameters.}

\item{x}{A numeric vector of data points or observations for which the log-likelihood will be calculated.}

\item{param}{A numeric vector of parameter values to evaluate against the log-likelihood function.}

\item{...}{Additional graphical parameters to customize the plot, such as `col`, `lwd`, etc.}
}
\value{
A list containing:
  \item{i}{Index of the parameter value that maximizes the log-likelihood.}
  \item{parami}{Parameter value that maximizes the log-likelihood.}
  \item{yi}{Maximum log-likelihood value corresponding to the maximum parameter.}
  \item{slope}{Slope values calculated around the maximum; "NA" if out of bounds (less than 1 or more than the number of parameter values).}
}
\description{
This function evaluates a log-likelihood function over a range of parameter values
and identifies the maximum likelihood estimate (MLE) of the parameters based on the provided data.
}
