% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nticketspackage.R
\name{ntickets}
\alias{ntickets}
\title{Calculate Optimal Number of Tickets to Sell for an Overbooked Flight}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of Seats on a flight.}

\item{gamma}{Confidence level (e.g., 0.05 for 95% confidence).}

\item{p}{Probability of a passenger showing up.}
}
\value{
A named list containing
\describe{
\item{nd}{the optimal number of tickets to sell based on discrete calculations.}
  \item{nc}{the optimal number of tickets to sell based on continuous calculations.}
  \item{N}{the number of seats available (input).}
  \item{p}{probability that the passenger shows up (input).}
  \item{gamma}{confidence level that the airline wants to achieve regarding the risk of overbooking (input).}
}
}
\description{
This function calculates the optimal number of tickets to sell for an overbooked flight using both discrete and continuous approximations.
It helps maximize revenue while minimizing the risk of overbooking.
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
