% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLAB3.R
\name{plot_tree}
\alias{plot_tree}
\title{Random Plotting in LAB3}
\usage{
plot_tree(data)
}
\arguments{
\item{data}{data gathered from spruce data frame}
}
\value{
A sophisticated plot
}
\description{
Random Plotting in LAB3
}
\examples{
plot_tree <- function(data)
}
